+++ src/components/common/assets.js
@@ -4,6 +4,7 @@ export const imageMap = {
     logo: `${assetFolder}/logo.png`,
     menuHome: `${assetFolder}/home.svg`,
     menuInsurance: `${assetFolder}/wallet.svg`,
+    menuAdditional: `${assetFolder}/social/skype.svg`,
     pdfIcon: `${assetFolder}/pdf.svg`
 };