+++ src/pages/index.js
@@ -4,6 +4,10 @@ import './insurance/tabs';
 import './insurance/banner';
 import './insurance/payment-form';
 import './insurance/calculator-form';
+import './additional';
+import './additional/game';
+import './additional/writer-form';
+import './additional/tabs';
 
 import './common/accountTable';
 import './home/account';